+++ src/RobotStudio.Simulation/MotionProfileShape.cs
@@ -1,0 +1,8 @@
+namespace RobotStudio.Simulation;
+
+public enum MotionProfileShape
+{
+    Stationary,
+    Triangular,
+    Trapezoidal
+}